layout update + remove unneeded cast

+++ cmd/game/main.go
@@ -15,7 +15,7 @@ const (
 
 func main() {
 	ebiten.SetWindowTitle(title)
-	ebiten.SetWindowSize(int(config.LayoutWidth*config.Scale), int(config.LayoutHeight*config.Scale))
+	ebiten.SetWindowSize(config.LayoutWidth*config.Scale, config.LayoutHeight*config.Scale)
 	theme.ApplyTheme()
 
 	cfg := &game.Config{
